requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
google-api-core>=2.15.0
